requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.4
python-dotenv==1.0.1
